Sorted the syllabi array, upper-cased the program, wrapped lists
Sorted the array contained the syllabi so that the the lists makes
much more sense now.

Transformed the $program variable to uppercase in the list headings,
which looks much nicer, and actually seems like it's acceptable.

Wrapped the lists in a div so that they can be made to sit nicely
next to one another without one title overtaking the other and pushing
the entire second list down. Dumb lists.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>